add Level-2 (storage of input)
