+++ .gitignore
@@ -18,3 +18,5 @@ text-ui-test/EXPECTED-UNIX.TXT
 
 # .class files
 *.class
+.vscode
+.idea
